pyproject.toml
[tool.poetry]
name = "fast-template-cassandra"
version = "0.1.0"
description = ""
authors = ["Ryan Jones <ryan.jones26@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_template_cassandra"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
requests = "^2.28.2"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
opentelemetry-distro = "^0.38b0"
opentelemetry-exporter-otlp-proto-http = "^1.17.0"
cassandra-driver = "^3.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
